flux_monitor/requirements.txt
 fastapi==0.111.0
 uvicorn[standard]==0.30.1
 celery[redis]==5.4.0
 redis==5.0.4

 SQLAlchemy==2.0.31
 asyncpg==0.29.0
 alembic==1.13.2

 pydantic-settings==2.4.0
 python-dotenv==1.0.1
 anyio==4.4.0

 playwright==1.46.0

 streamlit==1.37.0
 pandas==2.2.2
 psycopg2-binary==2.9.9
